+++ TODO
@@ -1,16 +1,21 @@
 Creer les groupes au moement de la creation de l'etude
-Interface supervisor passer le visit group en colonne
-Source du review manager a revoir !
-Reimplementer un send mail rapide pour visit upload manager ? appeler eme?
-Table spécifique review necessite d'ajouter modalite pour eviter collusion de nom, idem nom fichiers scpeficique
 
-technique : 
+Passage Visit Group : 
+=> Study  public function getSpecificGroup(String $groupModality) => Reste que le visit builder ...
+=> study getStudySpecificGroupManager => devrait aussi surement disparaitre => Reste statistics
+=> Table spécifique review necessite d'ajouter modalite pour eviter collusion de nom, idem nom fichiers scpeficique
+
+Refactoring : 
+Se repencher sur la logique du modele => quid des mails
 passe tout acces aux visites via visit group manager
 Le niveau patient et study iterent sur le group level
-=> Study  public function getSpecificGroup(String $groupModality) => Reste que le visit builder ...
-=> study getStudySpecificGroupManager => devrait aussi surement disparaitre => Reste statistics et reminder email controller
-Se repencher sur la logique du modele => quid des mails
-
 Fusionner le tree et l'API get review
+
++ Tard : 
+Reminder Visit Upload manager a revoir ? appeler eme?
+Review reminder a revoir
 Nb tree : si une visite review ajoute toute les autre groups visits en lecture
 A voir securité access review => dissocier niveau de reviewer ou jouer sur les job ?
+
+Tests:
+Source du review manager a revoir => A tester + verif que statistics ne casse pas trop
